[server] Move schemas into validation directory
